+++ acetime/zone_processor.go
@@ -1,7 +1,7 @@
 package acetime
 
 import (
-	"github.com/bxparks/AceTimeGo/zoneinfo"
+	"github.com/bxparks/acetimego/zoneinfo"
 	"strings"
 )
 
@@ -32,26 +32,25 @@ const (
 type zoneProcessor struct {
 	zoneInfo   *zoneinfo.ZoneInfo
 	year       int16
-	isFilled   bool
 	numMatches uint8
 	matches    [maxMatches]matchingEra
 	tstorage   transitionStorage
 }
 
 func (zp *zoneProcessor) isFilledForYear(year int16) bool {
-	return zp.isFilled && (year == zp.year)
+	return year == zp.year
 }
 
 // initForZoneInfo initializes the zoneProcessor for the given zoneInfo.
 func (zp *zoneProcessor) initForZoneInfo(zoneInfo *zoneinfo.ZoneInfo) {
 
 	zp.zoneInfo = zoneInfo
-	zp.isFilled = false
+	zp.year = InvalidYear
 }
 
 // Clear cache, used only for tests.
 func (zp *zoneProcessor) reset() {
-	zp.isFilled = false
+	zp.year = InvalidYear
 }
 
 func (zp *zoneProcessor) isLink() bool {
@@ -64,7 +63,6 @@ func (zp *zoneProcessor) initForYear(year int16) errType {
 	}
 
 	zp.year = year
-	zp.isFilled = true
 	zp.numMatches = 0
 	zp.tstorage.init()
 	if year < zp.zoneInfo.StartYear-1 || zp.zoneInfo.UntilYear < year {
@@ -673,7 +671,7 @@ type findResult struct {
 //
 // Adapted from ExtendedZoneProcessor::findByEpochSeconds(epochSeconds)
 // in the AceTime library and atc_processor_find_by_epoch_seconds() in the
-// AceTimeC library.
+// acetimec library.
 func (zp *zoneProcessor) findByEpochSeconds(epochSeconds ATime) findResult {
 	err := zp.initForEpochSeconds(epochSeconds)
 	if err != errOk {
@@ -707,7 +705,7 @@ func (zp *zoneProcessor) findByEpochSeconds(epochSeconds ATime) findResult {
 //
 // Adapted from ExtendedZoneProcessor::findByLocalDateTime(const LocalDatetime&)
 // in the AceTime library and atc_processor_find_by_local_date_time() in the
-// AceTimeC library.
+// acetimec library.
 func (zp *zoneProcessor) findByLocalDateTime(ldt *LocalDateTime) findResult {
 
 	err := zp.initForYear(ldt.Year)